Organize library modules into logical directories
Group the 20 loose .rs files from the tractor-core merge into
directories that reflect the architecture:

- glob/     — matching, pattern, normalized_path, files
- model/    — report, rule, tree_mode, normalized_xpath
- mutation/ — replace, xpath_upsert, declarative_set
- xot/      — builder, transform
- wasm/     — mod (bindings), ast (serialization types)
- languages/info.rs  — was language_info.rs
- output/source_utils.rs — was source_utils.rs

Backward-compatible pub use re-exports in lib.rs preserve all
existing crate:: and tractor:: import paths — zero changes needed
in binary-side code (cli/, executor/, format/, input/).

Result: 5 CLI-side loose files + 14 organized directories.
