+++ tractor/src/lib.rs
@@ -6,49 +6,88 @@
 //! - XPath 3.1 query execution
 //! - Multiple output formats
 
+// ---------------------------------------------------------------------------
+// Module structure
+// ---------------------------------------------------------------------------
+
+// Parsing & tree construction
 #[cfg(feature = "native")]
 pub mod parser;
-pub mod xpath;
-pub mod output;
-pub mod render;
-pub mod replace;
-pub mod source_utils;
-pub mod language_info;
-#[cfg(feature = "native")]
-pub mod files;
-pub mod xot_builder;
-pub mod xot_transform;
-#[cfg(feature = "native")]
-pub mod xpath_upsert;
-#[cfg(feature = "native")]
-pub mod declarative_set;
-pub mod normalized_xpath;
-pub mod normalized_path;
-pub mod glob_pattern;
-pub mod glob_match;
-pub mod report;
-pub mod rule;
-pub mod tree_mode;
-
-// Language transforms - available for both native and WASM
+pub mod xot;
 pub mod languages;
 
-// WASM support
+// Querying
+pub mod xpath;
+
+// Output & rendering
+pub mod output;
+pub mod render;
+
+// Core data types
+pub mod model;
+
+// File discovery & path handling
+pub mod glob;
+
+// Code mutation
+pub mod mutation;
+
+// WASM bindings
 #[cfg(feature = "wasm")]
 pub mod wasm;
+
+// ---------------------------------------------------------------------------
+// Backward-compatible re-exports
+//
+// These preserve all existing `crate::` and `tractor::` import paths so that
+// consumers (the binary crate, tests, etc.) don't need import changes.
+// ---------------------------------------------------------------------------
+
+// xot/ modules
+pub use xot::builder as xot_builder;
+pub use xot::transform as xot_transform;
+
+// languages/ additions
+pub use languages::info as language_info;
+
+// output/ additions
+pub use output::source_utils;
+
+// model/ modules
+pub use model::report;
+pub use model::rule;
+pub use model::tree_mode;
+pub use model::normalized_xpath;
+
+// glob/ modules
+pub use glob::matching as glob_match;
+pub use glob::pattern as glob_pattern;
+pub use glob::normalized_path;
+#[cfg(feature = "native")]
+pub use glob::files;
+
+// mutation/ modules
+pub use mutation::replace;
+#[cfg(feature = "native")]
+pub use mutation::xpath_upsert;
+#[cfg(feature = "native")]
+pub use mutation::declarative_set;
+
+// wasm/ additions
 #[cfg(feature = "wasm")]
-pub mod wasm_ast;
+pub use wasm::ast as wasm_ast;
+
+// ---------------------------------------------------------------------------
+// Flat re-exports (convenience API)
+// ---------------------------------------------------------------------------
 
 #[cfg(feature = "native")]
 pub use parser::{
     detect_language, SUPPORTED_LANGUAGES,
-    // Unified parsing pipeline (always returns Documents)
     parse_to_documents, parse_string_to_documents,
     load_xml_string_to_documents, load_xml_file_to_documents,
     XeeParseResult,
-    // Version info
     get_language_abi_versions, LanguageAbiInfo,
-    // Timing stats
     print_parse_timing_stats,
 };
 pub use xpath::{XPathEngine, Match, XmlNode, print_timing_stats, Documents, DocumentHandle};
