+++ tractor/src/model/mod.rs
@@ -1,0 +1,6 @@
+//! Core data types: reports, rules, tree modes, and normalized expressions.
+
+pub mod report;
+pub mod rule;
+pub mod tree_mode;
+pub mod normalized_xpath;
