+++ tractor/src/xot/mod.rs
@@ -1,0 +1,4 @@
+//! Xot XML tree building and transformation infrastructure.
+
+pub mod builder;
+pub mod transform;
